buildpkg: don't use full pathnames by default.
Previously, buildpkg would use /usr/bin/gcc and /usr/bin/g++ as CC and CXX if no
path was specified.  This was needed to work because if the scripts in
$HOME/bin/build were passed unadorned program names, they would simply end up
rerunning themselves (because $HOME/bin/build is first in the $PATH).

Now, we only add $HOME/bin/build to the $PATH if we are passed a compiler
argument on the command line.  Otherwise, we simply use gcc and g++.  This
avoids a problem with libtool, which seems to break rather spectacularly if CC
(respectively CXX) is a full path.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.ath.cx>