vimrc: obscure vim 7 function call.
The Vim 6.3 used in CentOS 4 wants braces around a function name with a hash
sign in it, like pathogen#infect.  However, newer Vims don't want that and
complain if you do.  This is true regardless of whether the function is ever
actually called.  Since pathogen is unlikely to work in such a case anyway,
create a funcref to pathogen#infect and call it that way to obscure it from old
versions of vim.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>