vimrc: force-load pathogen.
Vim 7.0 on CentOS 5 seems to have a serious issue with autoloading, causing
warnings about loading pathogen every time.  Avoid this problem by using a
runtime command to help it out.  This command (or an equivalent) would have been
run anyway on autoload, so it is a no-op on properly-configured Vims.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>