vimrc: set comments to exactly n:>.
The default comment setting includes the C settings, which cause bulleted lists
in emails to have the * on successive lines.  This is undesirable, but previous
attempts at fixing this by setting the comments option to empty were also
undesirable because quoting did not work correctly.  Fix this by setting the
comments option to exactly n:>, so that quoting works and so do bulleted lists,
which are then handled by the formatoptions option.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>