vimrc: improve whitespace handling.
The previous techniques for handling trailing whitespace were inadequate.
Different file types need different handling; for example, diffs should be
expected to have a space at the beginning of the line if the line is empty.
Also, the insert and normal mode patterns are very similar and can be generated
from each other.  Finally, default trailing whitespace detection to off for help
files and allow toggling this with <Leader>t.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>