prompt: ensure text displays properly in old zsh.
In old versions of zsh, the %V prompt character does not exist.  As a
consequence, the p_mode (ok, no, ◈◈) is never displayed.  Since these old
versions do not support zle-keymap-select, there is no way to ever indicate that
the prompt is in vi-cmd mode, so the best we can do is to make the "ok" or "no"
show up properly.  Modify p_mode to use a syntax that makes this possible.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>