vimrc: prevent double whitespace handling.
If a window were split, and one entered insert mode on the bottom, the normal
mode whitespace handling would occur as well, resulting in spaces that had just
been typed being flagged as errors.  Fix this by using WinEnter and WinLeave
in addition to BufWinEnter and BufWinLeave.  This has the side effect that only
the current buffer is highlighted for trailing whitespace, but that's not a huge
problem.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>