gitconfig: add a setup alias.
Oftentimes, I want to create a temporary repository, either for building
patches against a tarball, or to use git grep or other git tools on the
repository.  Create an alias, setup, that turns the current directory into a
git repository, checking in all the files and tagging the initial commit as
"start".

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>