gitconfig: use a fixed time for git setup.
If a fixed time is used for git setup, applying identical patches on top of
the repository will result in identical commits, and the SHA-1 values will
not change, which can help reduce diff noise when checking those patches in.
Use a fixed time (my birthday) to make this happen.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>