zshrc: prevent blank lines when setting tabs.
tabs outputs some control characters which move the line to the right, and
then outputs a newline.  This results in a blank line every time a new
terminal window is opened.  Convert this newline to a carriage return, so
that no blank line is output and the prompt starts at the first line and
column.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>