vimrc: enable syntax highlighting earlier.
A recent change moved the autocommands before the syntax highlighting
was enabled to avoid a problem where :hi link caused errors due to the
lack of the Error type being defined.

However, if the syntax highlighting is enabled after the autocommands,
anytime we want to override syntax settings, they're blown away.  This
breaks the AsciiDoc syntax highlighting override, which makes editing
AsciiDoc slow.  Retain the current position for the colorscheme to
not break the change described in the first paragraph, but move the
:syntax enable command before the autocommands to fix the latter issue.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>