vimrc: disable ttyfast for SSH.
Vim automatically sets ttyfast if the TERM type is xterm (or,
apparently, xterm-256color).  However, it's still set even over SSH
connections, which can be very slow (or have high latency) in some
cases.  Clear this variable if an SSH connection is detected, so that
less data is sent over the pipe.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>