zshrc: fix completion with libpam-tmpdir.
libpam-tmpdir creates a /tmp/user directory with root:root 711
permissions and then creates 700 directories for each user (owned by the
user) inside of it.  The name of each user directory is the user's
numeric UID.

zsh by default attempts to complete all components of a path, but this
prevents completion from working inside an unreadable directory.  Enable
the accept-exact-dirs option, which prevents attempting to complete
prefixes that match an existing directory, to fix this problem.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>