zshrc: clean up local::lib variables.
local::lib will try to insert undesired environment variables when
parsing /etc/profile.d, which is executed as part of zshrc on Red Hat
and CentOS systems.  This leads to attempts to install Perl modules
installing to $HOME/perl5, which may or may not be in the load path,
which causes depending Perl modules to fail to install.

Since this behavior is so broken and so undesirable, ensure that these
variables are removed from the environment and that PATH is sanitized to
remove $HOME/perl5/bin.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>