vimrc: load Perl refactoring scripts from $HOME.
It's useful to have a place for Perl refactoring scripts, which could be
in a module somewhere under $HOME/lib/perl5.  Load that path into @INC
if it's a directory and Perl is available in Vim.  Don't invoke the
:perl command unless that directory exists, as some systems dynamically
load Perl but don't depend on it, and invoking it unconditionally can
cause segfaults.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>