zshrc: improve robustness of 256-color terminal detection
The detection of 256-color terminals matched any data at the end to deal
with a possible " (deleted)" that may occur.  Instead of simply placing
a glob at the end, look specifically for the "(deleted)" text, and
remove it and any spaces.  Space removal and "(deleted)" detection are
done separately and not anchored because CloudLinux prepends
"(deleted) " instead of appending " (deleted)".

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>