zsh: prefer vimx to vim.
On RHEL systems, vim is always a version compiled without graphical
support, and hence a version without X copy and paste, even if run
inside a graphical environment.  However, vimx, if available, has such
functionality.  On Debian systems, vimx does not exist, but vim is
generally the same binary as gvim, if installed, and so it has that
support.  Prefer vimx to vim to ensure that we select a version with
proper copy and paste support if available.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>