zsh: print branches even for root.
The original idea was that root would not need to work with git and that
not printing branches would discourage accidentally using git and
creating files unreadable by a normal user (and thus a broken
repository).  However, this ended up being a bad idea for system
administration reasons, where deploying code via git is desirable and
may need to be done as root.  The code that was present also had the
problem of not printing the branch name or the opening parenthesis,
while the close parenthesis was present.  Print branch names even for
root.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>