gpg.conf: ensure agent is usable at first login.
Due to Debian bug #790316, if "use-agent" is present in gpg.conf,
/etc/X11/Xsession will start gpg-agent.  However, the
DBUS_SESSION_BUS_ADDRESS environment variable is not set at that point;
it only becomes set when the final exec into the session occurs (due to
the use of dbus-launch).  As a result, the pinentry spawned cannot
connect via libsecret and fails.

However, .zshrc starts the agent automatically if it is not running.
Remove the "use-agent" item, which is automatically enabled for GnuPG 2
anyway, so that it will be started by zsh, at which time the right
environment will be set up.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>