prompt: avoid power checking on slow boxes.
On slow devices such as Android phones, power checking is either not
useful because the device is wired, or in the case of Android, not
available because the kernel doesn't expose it in a way we can use.
Regardless, it's very slow to perform all the checking required.
Disable that part of the prompt if ~/.slowbox is set.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>