zsh: paper over editor inconsistencies.
On Red Hat-based systems, the console Vim version which is built with X
clipboard support is called vimx; vim is a basic version, and vi is a
tiny version of Vim.  Debian defaults vim to the most capable version,
and vi invokes that same Vim, using the alternatives system.

Clearly Debian's result is better, as it produces a more consistent
experience, and it allows one to install exactly one binary that meets
all their needs.  However, Red Hat is not likely to change this in
existing deployments, meaning that it has to be worked around.

Despite the namespace pollution that occurs doing so, writing a shell
function to deal with this is the best way to handle it.  Call such a
function bmc_editor, and teach it how to determine the best graphical
and console editors available.  Adjust the remainder of the places in
.zshrc and .zshenv to call this function.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>