zshenv: invoke gvim directly.
Many years ago, before the git history of this repository, there was
some program that did not invoke $EDITOR in a shell, resulting in
breakage when used with "gvim -f", as that string does not name a valid
executable.  A wrapper script, called "gvim-f", was used to paper over
this issue.  In the past nine years, the program (whose name I don't
remember) has probably been fixed, either directly or by invoking
sensible-editor.  Invoke gvim directly.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>