prompt_bmc_setup: conditionalize hostname coloration
Some older versions of zsh have a bug which causes them to produce odd
coloring during tab completion.  Before the hostname colorization, this
caused the text to be printed in light blue, which is readable.
Afterwards, it was in dark purple.  Disable this change on older zsh so
that the text is at least legible.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>