zshrc: don't alias vim to vi.
If vi was installed, but not vim (say, on a CentOS system without
vim-enhanced), "vim" would end up being an alias to vi.  This results in
breakage if a real Vim is later installed, and produces odd behavior
nevertheless.  Prevent this by not creating a vim alias if $EDITOR is
set to vi.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>