zshenv: disable rogue profile scripts
CentOS ships with several "helpful" scripts in /etc/profile.d.
Unfortunately, one of these scripts uses the variable TMP and then
clears it unconditionally.  This breaks configuration that wants to set
this value to a usable temporary directory.

This, among other unhelpful behavior from random scripts in
/etc/profile.d, demonstrates that it's safer to avoid the OS's
configuration.  Unset the globalrcs option as soon as we can.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>