zshenv: use zstat instead of stat
There isn't a portable way to use stat, as the custom format option
differs between GNU userlands and FreeBSD-based userlands (including
Darwin).  Such a utility might not even exist.

However, the zstat builtin can be used for that purpose, and it provides
the information we want.  Use it instead of stat.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>