vimrc: remap Ctrl-A to Ctrl-S.
Ctrl-A is the prefix key for tmux, and it's easy to type Ctrl-A when I
meant to type Ctrl-W to switch windows.  Use Ctrl-S, as it's normally
unmapped, and unmap Ctrl-A instead to prevent thinkos.  That it is used
for flow control should not be relevant as my zsh configuration disables
flow control, and I rarely use the functionality, so skipping it in the
rare case when I'm not using zsh should be fine.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>