zsh/prompt_bmc: restore hostname coloring
In zsh 5.3, the odd coloring that occurred in the command during tab
completion has gone away.  In earlier versions when hostname coloring
was not used, this was a light blue, which, while odd, was easy enough
to see.  However, with hostname coloring, this became dark purple or
grey, which was illegible.

Since zsh 5.3 has fixed this issue, restore the hostname coloring that
was available previously if we have such a version.