zsh/prompt_bmc: fix highlighting as root
zsh 5.3 highlights the text on the command line to be the same color as
the last color in use after stopping for tab completion.  As root, this
is red, due to the "#".  To work around this, print the final space as
white.  This won't affect rendering, but it will make the tab completion
use the correct color.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>