vimrc: flag all perlcritic errors in a theme
If we're using a given theme, we want to flag all errors in that theme;
otherwise, we wouldn't be using that theme.  Add the "-1" option to
Syntastic's perlcritic arguments so that we get all possible errors
specified in the theme in question.  Note that "-1" is only used if a
theme is set; otherwise, the default behavior is preserved.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>