vimrc: set filetype for *.h to "c"
Most of my development is done on projects where *.h reflects a C file.
There are other widely accepted extensions to indicate that the header
is for C++, such as .hh, .hpp, or (on Unix) .H.  Setting the file type
to C fixes warnings due to parameters named "template" and the like when
working on Git.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>