vimrc: set Ctrl-W as termkey
The default behavior if no termkey is set is to have Ctrl-W perform
certain special actions.  However, in that case, repeating Ctrl-W
doesn't send a Ctrl-W to the terminal.  Set termkey to Ctrl-W so that
doubling it does send that key to the terminal window.  This behavior is
consistent with tmux's.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>