Remove gvim-f.
A long time ago, there was a package somewhere in Debian that did not
invoke the shell to handle $EDITOR (and it did not use sensible-editor,
which does).  This led to breakage when using "gvim -f" as $EDITOR.
However, I'm not aware of any package that still has this problem and
I've been using "gvim -f" as $EDITOR for some time, so remove this
helper.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>