zshrc: improve rename alias
Debian ships a separate package called "rename" which provides a binary
of "file-rename".  This binary is the same Perl rename as "prename", but
the latter is likely to be discontinued and causes a warning to be
printed.  Fix the alias so we prefer file-rename if present and fall
back to prename if the former is not available (say, on Red Hat
systems).

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>