vimrc: disable colors for ex mode
When in ex mode, vim does not use syntax highlighting to display text.
However, it uses the default syntax highlighting colors to display the
background and prompt, which doesn't really add much but makes the
terminal background unhelpfully opaque.

Detect the mode we're starting up in by checking the mode function in
.vimrc.  If it is “ce” (normal ex mode) or “cv” (Vim's ex mode), then
set a file-local variable indicating that and turn off syntax
highlighting.  We turn off whitespace handling as well as that depends
on having syntax highlighting set up.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>