zsh/prompt_bmc: fix overly large battery values
On some laptops, the kernel can accidentally think that the amount of
charge available is in excess of the current maximum capacity of the
battery.  Therefore the charge can be "101%".  In this situation, clamp
the value to 100%, as the thing we really care about is whether the
battery is fully charged.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>