zshrc: improve gvim and gview wrapper functions
Add a -z option to create a new gvim instance with a certain server
name; the existing -t option can be used to open additional files in the
same instance.

Set a default server name if we're using X forwarding, as sometimes this
doesn't get set properly when the Vim instance is running on a CentOS
system.  Note that this default works exactly the way that Vim does when
it does set the default server name.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>