vim: don't disable highlighting in GUI
If TERM is "dumb", that can be because we're in a dumb terminal, or it
can be because we're in the GUI and we were started from a place that
doesn't have a terminal.  If we're in the GUI, we still want to enable
highlighting regardless of our terminal situation, so only check the
terminal type if the GUI is not running.

Update the ct-colors repository to a commit that includes a similar
change.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>