zsh/prompt: handle terminals with many colors correctly
The code to check for using fancy colors was checking for the number of
colors to be equal to 256.  However, there are now terminal definitions
that support a very large number of colors, and we want to use the same
palette there as well.  Check that the number of colors is at least,
instead of exactly, 256 colors.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>