zsh: add full true color support
Update the color handling code for the prompt, including autoloaded
shell functions, to accept full 24-bit color values in true color mode.
Use setaf and setab if they're available in the terminal definitions and
the terminal indicates support for more than 256 colors; otherwise,
generate the escape sequences ourselves.  Note that we use the escape
sequence form with semicolons instead of colons, as this form is more
widely accepted.

Add a helper function, color_count_for, that can perform an action based
on whether the terminal is true color, 256 color, 16 color, or black and
white.

Use the new code for less highlighting as well.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>