zsh: move palette selection into its own function
Instead of keeping the prompt palette selection in the prompt code, move
it out into a separate function which can be used additionally to look
up colors for less and grep.  Any piece of code can now simply call a
shell function to get a proper palette value for the current terminal.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>