zsh: do not set all prompts to bold
The color_fg function expects an empty or non-empty value to disable or
enable bold, but the bmc_palette function uses "yes" and "no".  This
causes the bold attribute to always be set, as both values are
non-empty.  Adjust this code to only emit the bold field if it's set to
"yes".

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>