zsh: fix rendering of prompt in subshells
When entering a subshell and typing Ctrl-L, the "ok" or "no" portion of
the prompt would contain the shell level information instead.  This
occurred because the syntax used to unset the entries in psvar not only
unsets those entries, but deletes them from the array, causing later
entries in the array to be shifted left.  This was not previously a
problem until psvar[4] was used because there was nothing to shift into
the psvar[3] slot.

Since all the code that handles these variables also handles empty
variables in the same way as unset ones, avoid this problem by using an
empty string instead.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>