gitconfig: add branch mark aliases
Oftentimes, when working with many topic branches which are based on top
of one another, it makes sense to indicate the base of the branch in
order to allow easy rebasing onto an updated version.  Introduce two
aliases to work with branch marks, which are empty commits at the base
of a branch that have a message containing only a colon and a branch
name.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>