gitconfig: anchor match in mark alias
The mark alias lacked any anchor for the match, meaning that if one ran
"git mark abc" and a mark for "abc-fixup" were present, the wrong mark
could be chosen.  Since Git always ensures that a commit message ends
with a newline, use a newline at the end of the match to avoid any
ambiguity.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>