zsh: use ps to detect parent processes
On Darwin, there is no /proc filesystem to query.  The easiest way to
find out the process name on both Linux and Darwin is to use "ps".
However, on Linux, tmux server processes show up as "tmux: server", so
ensure we match them correctly by expanding the case statements that
match them.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>