zshenv: remove setting of GIT_ASKPASS
In previous contexts, Git over HTTPS was used only for public clones and
private clones would not use HTTPS, making it preferable to simply fail
when Git prompted for a username and password.  However, now that HTTPS
is used for a wide variety of operations, remove the setting of
GIT_ASKPASS.

In addition, the git-lfs testsuite fails with this environment variable
set, making it even more unhelpful.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>