bin: add git-at
Git provides an ability to find a revision that matches a regular
expression from the current branch.  However, the syntax to do so is
very difficult to type, requiring both the specification of a revision,
and braces, essentially requiring that the expression be quoted.

Add a wrapper for this syntax, git at, which can look for a regex in a
commit message, defaulting to HEAD, and optionally only in the commit
summary.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>