Revert "git/gitconfig: always enable reflogs"
The always option to core.logAllRefUpdates is fairly new and not
supported in Debian stable. In addition, libgit2 only recently learned
about it and it isn't supported in any released version. Consequently,
tools like cargo that use libgit2 under the hood are unable to clone
repositories due to this setting.

This reverts commit 1a0563f76b4ef64186e746fc25f70b0b3b4e173c.