vim/vimrc: autodetect LSP root
One of the submodules we use is ALE, which provides Language Server
Protocol support. There's an autodetection mechanism built in, but in
general, we want to detect the project root as the top of the Git
repository. Add some functions to set per-buffer variables so the LSP
root can be detected.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>