zsh/zshenv: replace echo and printf with here-strings
On Windows Subsystem for Linux 1809, zsh hangs if given a pipe inside a
command substitution. In several cases, it's possible to avoid these by
using a here-string instead, so do so.

In the remaining case left in this file, the code will not be executed
except if a ~/.gnupg/gpg-agent.conf setting is available, so simply
omitting that file should be sufficient to make things work.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>