zsh/functions: don't use hostcolor on WSL
On Windows Subsystem for Linux 1809, zsh hangs if given a pipe inside a
command substitution. In the palette code, it's not possible to avoid
these, since we want to hash the hostname without a trailing newline,
and here-strings automatically include a trailing newline.

If /proc/version exists and it looks like it's from a WSL instance, set
the host color to white instead.

Additionally, port over some of the code for working systems to use
here-strings, since they are more compact and easier to read.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>