zsh/zshrc: don't set GREP_COLORS on WSL
On Windows Subsystem for Linux 1809, zsh hangs if given a pipe inside a
command substitution. This is true even if the pipe is in a shell
function and the shell function is called from the command substitution.

If /proc/version exists and it looks like it's from a WSL instance,
don't set GREP_COLORS, since the shell function used invokes a pipe to
perl.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>