vim/vimrc: avoid flagging camel case words as misspelled
When using spell checking for comments in certain languages, it's common
for camel case words, such as class or method names, to be flagged as
misspelled. This is quite distracting for the user. Adding words to the
dictionary would bloat the dictionary significantly and would need to be
done for every project, so add a function which can be used to ignore
words with capital letters in them when spell checking programming
languages.

Note that this function would not be suitable for general-purpose text,
only programming language comments.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>