vimrc: disable modelines
Modelines in Vim are implemented with a blacklist, which is bad for
security. We use the securemodelines plugin, which uses a whitelist, so
the modeline setting is no longer needed for functionality. The
securemodelines plugin disables it for us, but just in case, disable the
modeline setting for security reasons.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>