vimrc: don't modify files on save
When working in a project with an EditorConfig file, if that file
directs that trailing whitespace be stripped, the EditorConfig plugin
will do that automatically on save. There are some cases in which it's
useful to insert trailing whitespace temporarily for test purposes, and
Vim highlights these as errors anyway. Disable the automatic stripping
on save.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>