vimrc: switch go LSP tool from bingo to gopls
Bingo is no longer being developed in favor of a modified version of
gopls.  In addition, the former had some cases in which it would
falsely flag items; the latter does not have this issue.  Since ALE
requires us to explicitly configure entries for Go, add gopls and remove
bingo.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>