bin: add a script to unwrap text
Oftentimes, one wants to write text in a wrapped format (such as in a
commit message) but enter it into a system which normally does wrapping
on its own (such as the GitHub web interface).  Add a script,
format-text, which preserves double newlines as paragraph markers but
unwraps paragraphs into single lines.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>