bin: add a script to format commit message bodies
When writing a commit message, it's good form to wrap one's lines at 72
characters.  However, the GitHub web interface preserves newlines, so if
a commit message is entered there, these newlines will be preserved,
leading to ugly formatting.  Add a script to wrap the output the body
(but not the summary) of a given commit using format-text so it can be
pasted nicely into the GitHub web interface.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>