vimrc: use clippy with ALE when available
By default, ALE uses cargo check to check Rust code.  However, if
enabled, it's possible to use cargo clippy, which not only checks for
the standard errors and warnings, but other style checks as well.  Since
there's really no downside to this, enable it if we're using rustup and
therefore have it available.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>