zsh/zshenv: allow global RC settings
On some GitHub systems, there are useful settings such as PATH entries
that we'd like to have from the global RC files.  In addition, we're not
using a Red Hat or derivative system, so we don't have to worry about it
overwriting various environment variables by failing to use local or
using unset.

Remove the unsetting of the globalrcs option.

Signed-off-by: brian m. carlson <bk2204@github.com>