vimrc: swap mappings for "* and "+
On most systems, the "* and "+ registers both refer to the system
clipboard.  On X11, however, the former refers to the PRIMARY selection
(the one chosen by highlighting with the mouse) and the latter refers to
the CLIPBOARD selection (the one usually accessed with Ctrl-C and
Ctrl-V).

It's obvious that a double backslash is easier to type than a backslash
and C, so the former should be used for the more common selection.
Because the CLIPBOARD selection is easier to access with the keyboard,
which is faster and more precise than using the middle button to paste,
let's switch the mappings such that "+ maps to the double leader key.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>