Makefile: install links for gems
Instead of overriding GEM_PATH, let's simply use the default value,
which installs files into ~/.gem.  However, we want to store them in
.local/share/gems, so let's create a symlink instead.  This avoids us
setting a GEM_PATH which may or may not work on various systems.

When we create symlinks, remove the destination beforehand so that ln
does not dereference the destination if it points to a directory.  Note
that there is an option in GNU ln, -n, which implements this behavior
for us, and it is implemented as well in FreeBSD and macOS, but it is
not in POSIX, so use the rm default.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>