zsh: distinguish VISUAL and EDITOR
There is a distinction between VISUAL and EDITOR: the former is designed
for terminals which have an addressable cursor and the latter is for
dumb terminals if both are set.  In general, most Unix programs honor
this distinction, and Debian requires sensible editor behavior.

Since both are set here, let's set EDITOR to ex, which works even on
dumb terminals, provided a recent version of Vim (8.0.1454 or newer) is
being used, unless we're using a graphical interface.  Set VISUAL to a
more capable editor as before.  Don't assign values from one to the
other.

In addition, don't distinguish between shells in screen and tmux
sessions that have screen or tmux as their parent and ones that do not,
since otherwise starting a new interactive shell changes the editor,
which is unexpected and bizarre.  Only change the editor in an SSH
session, where obviously we will not have a graphical interface.

Because this behavior is tricky and there are a lot of edge cases, let's
add some tests using RSpec.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>