git: name default branch "dev"
There has been much discussion lately about the term "master" for Git's
default branch and its relationship to the problematic master/slave
terminology used in tech.  It's obvious in retrospect that the name is
suboptimal and if we were making the decision today, we'd pick a
different name.  Since there's now an option, there's no reason not to
make a better decision going forward.

Choose the name "dev" for the default branch for several reasons:

* It's short, easy to remember, and easy to type.
* It reflects the purpose for my projects, which is the main development
  branch.
* "Development" has a cognate in French and is not related in Spanish.
  "main", another popular choice, means something different and
  unrelated in French and will be confusing to some non-English speakers.
* It's not the branch name used by another major workflow.  "develop"
  would be confusing to people familiar with Git Flow.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>