mutt: place quotes around name in From header
RFC 5322 requires that the portion of the From field corresponding to a
human name (a display-name) be either one or more atoms or
quoted-strings.  An atom is not permitted to have dots, so for a name
with a dot after the middle initial, it's required to quote the display
name.

Until recently, Debian's version of mutt did this for us by rewriting
the address field.  However, it no longer does, and as a consequence, we
must produce a syntactically valid From header ourselves, so let's do
so.