zshenv: set a stack size limit
In recent versions of Linux, the value set for RLIMIT_STACK also affects
ARG_MAX.  If this value is unlimited, ARG_MAX is set to
4611686018427387903, which is much larger than the available amount of
memory in the system.  That means that a program trying to malloc that
amount of memory will fail.  pgrep falls into this category.

To help out these programs, let's set the stack size to 64 MiB, which is
quite generous, which in turn causes ARG_MAX to be 16 MiB.  This is a
sane and reasonable amount of memory a program might want to malloc, so
various programs will work as normal.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>