bin/format-text: handle lists and indentation gracefully
It can often be useful to include indented text, such as in a commit
message, to indicate that some text is literal.  We don't want to wrap
such text.

However, we also want to handle lists where we do want wrapping.  Let's
expand format-text to handle both these cases and format them nicely.
Since this is tricky, let's add some specs for this case as well.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>