vimrc: add a check for lsproot
Normally we set the project root for LSPs to the top of the Git
repository.  However, sometimes there are multiple projects in a
repository and we need some way to indicate that the LSP should use a
subdirectory as the directory.  Let's look for a file that's called
".lsproot" that indicates this.  This file can be ignored in
.git/info/exclude.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>