vimrc: adjusst settings for neovim-gtk
In case we're using neovim-gtk, let's set the GUI settings
appropriately.  Neovim does things a little differently and we won't
know that we're in GUI mode until after the .vimrc has already been
read, so let's set up a UIEnter autocommand to trigger when the GUI is
loaded and set the proper settings.

However, the UIEnter command also happens to be invoked when we're using
a TTY, and we don't want to adjust the number of lines or columns when
using a TTY.  Detect that we're using a GUI because we're using Neovim,
our standard output isn't a TTY, and we're in vi mode.  The latter is
required because if we're at the command line but really lack a TTY
(e.g., with a dumb terminal), we'll be in ex mode and won't want to try
to modify the terminal settings.

Because this is all a little complex, let's move the code to a function,
and move the GUI code to the bottom of the file so that the function is
defined before using it.  The top portion of the file must function even
when we're in a tiny version of Vim with no scripting support, so we
cannot define the function there.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>