bin: add a tool to work with the clipboard
There are a variety of tools to work with the clipboard that vary based
on the platform.  These tools have varying interfaces and may or may not
be installed.

Add a tool which abstracts these other tools to copy and paste from the
clipboard.  In addition, add support for a shell command to filter the
clipboard for ease of use.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>