zsh/zshrc: add better support for more terminal types
Right now, we force the use of xterm-*, even when vte-* is available,
which means that newer versions of terminfo can break terminal
definitions.  Let's use the proper terminal type so that we always have
a working terminal definition.

In addition, support several of the *-direct terminal types instead of
using *-256color and shoveling the true color terminal sequences over
the pipe while sticking our fingers in our ears and saying "la la la la
la!"  This will lead to more software that is functional out of the box
and generally less surprising behavior.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>