Revert "zshenv: set a stack size limit"
This limit was necessary when using some procps tools, since having an
unlimited value made the programs attempt to allocate absurd amounts of
memory.  However, this has been fixed in newer versions of procps and
shouldn't affect any stable Debian versions, so it's no longer needed.

Additionally, since this value is too high for WSL 1 environments,
having this set causes WSL to scream every time it loads a new terminal,
which is not great.  Let's revert this change and leave the value at the
defaults.

This reverts commit 45d2ddc541dcbad5a17133a9b6c65a954de4820a.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>