spec: test fallback of editor configuration
Let's test the fallback of editor configuration to make sure that we get
things in the proper order.  To do so, add a check for an environment
variable in zshenv that controls whether to set the PATH or not and then
use it in the spec code while setting a restricted PATH containing only
certain programs.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>