zsh/zshenv: switch to table-driven selection for editor
Currently, the editor code path is a giant if-elif-else block that
sometimes doesn't check for the binary to exist before setting an
editor.  This can cause various problems, such as setting gvim when it
isn't installed.

Let's fix this by switching to a table-driven approach which looks for
each editor in turn and explicitly checks if the program is present.
This way, we have the selection logic centralized in one place and can
easily determine the right editor automatically.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>