zsh/prompt: find AC and battery regardless of name
On some Linux systems, the AC device is not called "AC" in sysfs, and
presumably the battery need not be called "BAT0".  Let's iterate over
those devices and find out if one of them looks like a wired power
connection (that is, type Mains) and if one of them looks like a battery
of the right type.

Note that we ignore some types of batteries, because an external
Bluetooth device, such as a cell phone, can report its batter under this
class as well, and we're not interested in the batteries of those
devices.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>